requirements.txt
requests
fusepy==2.0.4
beautifulsoup4==4.5.1
